Make MeshBuilder compile in UWP project
