+++ Source/HelixToolkit.SharpDX.Shared/HelixToolkit.SharpDX.Shared.projitems
@@ -9,5 +9,18 @@
     <Import_RootNamespace>HelixToolkit.SharpDX.Shared</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)Core\Color4Collection.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Core\ExposedArrayList.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Core\IntCollection.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Core\Polygon.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Core\Polygon3D.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Core\Vector2Collection.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Core\Vector3Collection.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensions\CollectionExtensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensions\Vector3DExtensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Model\Geometry\Geometry3D.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Model\Geometry\MeshGeometry3D.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\TokenizerHelper.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utilities\TypeConverter.cs" />
   </ItemGroup>
 </Project>
