+++ Source/HelixToolkit.UWP/project.json
@@ -6,7 +6,8 @@
     "SharpDX.Direct2D1": "3.0.1",
     "SharpDX.Direct3D11": "3.0.1",
     "SharpDX.DXGI": "3.0.1",
-    "SharpDX.Mathematics": "3.0.1"
+    "SharpDX.Mathematics": "3.0.1",
+    "System.ComponentModel.TypeConverter": "4.3.0"
   },
   "frameworks": {
     "uap10.0": {}
